Enum term_painter::Attr

pub enum Attr {
    Plain,
    Bold,
    Dim,
    Underline,
    Blink,
    Reverse,
    Secure,
}

Lists possible attributes. It implements ToStyle so it's possible to call ToStyle's methods directly on a Attr variant like:

println!("{}", Attr::Bold.fg(Color::Red).paint("Red and bold"));

It is not guaranteed that the local terminal supports all of those formatting options. As already mentioned in the module documentation, you should use term directly to check the terminal's capabilities.

For more information about enum variants, see term::Attr Documentation.

Variants
Plain

Just default style

BoldDimUnderlineBlinkReverseSecure