Enum term_painter::Color

pub enum Color {
    NotSet,
    Black,
    Red,
    Green,
    Yellow,
    Blue,
    Magenta,
    Cyan,
    White,
    BrightBlack,
    BrightRed,
    BrightGreen,
    BrightYellow,
    BrightBlue,
    BrightMagenta,
    BrightCyan,
    BrightWhite,
    Custom(u16),
}

Lists all possible Colors. It implements ToStyle so it's possible to call ToStyle's methods directly on a Color variant like:

println!("{}", Color::Red.bold().paint("Red and bold"));

It is not guaranteed that the local terminal supports all of those colors. As already mentioned in the module documentation, you should use term directly to check the terminal's capabilities.

Note: Using Color::NotSet will not reset the color to the default terminal color.

Variants
NotSetBlackRedGreenYellowBlueMagentaCyanWhiteBrightBlackBrightRedBrightGreenBrightYellowBrightBlueBrightMagentaBrightCyanBrightWhiteCustom(u16)