Struct term_painter::Painted

pub struct Painted<T> { /* fields omitted */ }

Wraps an object of type T and a style. When attempting to print it, the given style is applied before printing and reset afterwards. All formatting traits (Display, Debug, ...) that are implemented for T are also implemented the wrapper type Painted<T>.