Crate term_painter

This is a crate for coloring and formatting terminal output. Simple example:

extern crate term_painter;

use term_painter::ToStyle;
use term_painter::Color::*;
use term_painter::Attr::*;

fn main() {
    println!("{} or {} or {}",
        Red.paint("Red"),
        Bold.paint("Bold"),
        Red.bold().paint("Both!"));
}

How to use it

Formatting works in two steps mainly:

Creating a style

Use this style to "paint" something and reviece a Painted object

1. Creating a style

To create a style a startpoint is needed: This can either be a startpoint with an attached modifier (like Red: modifies the fg-color) or the Plain startpoint, which does not modify anything. After that the startpoint can be modified by modifiers like bold() or fg().

extern crate term_painter;

use term_painter::ToStyle;
use term_painter::Color::*;
use term_painter::Attr::*;

fn main() {
    let x = 5;

    // These two are equivalent
    println!("{}", x);
    println!("{}", Plain.paint(&x));

    // These two are equivalent, too
    println!("{}", Red.paint(&x));
    println!("{}", Plain.fg(Red).paint(&x));
}

You can chain as many modifier as you want. Every modifier overrides preceding modifier:

println("{}", Plain.fg(Red).fg(Blue).paint("Apple")); // blue, not red

2. Use the style

After building the style, you can call paint to use it on some object. paint will return the wrapper object Painted that holds your object and the specified style. Painted implements Display and/or Debug if the type of the given Object, T, does. So the Painted object can be printed via println! or similar macros. When it gets printed, it will apply the given style before printing the object of type T and will reset the style after printing.

Note: paint will consume the passed object. This is no problem when passing constant literals (like paint("cheesecake")) or types that are Copy. Otherwise it could be confusing because just printing should not consume a variable. To prevent consuming, just pass a borrow to the object (with &). Example:

extern crate term_painter;

use term_painter::ToStyle;
use term_painter::Color::*;
use term_painter::Attr::*;

fn main() {
    let non_copy = "cake".to_string();  // String is *not* Copy
    let copy = 27;  // usize/isize *is* Copy

    println!("{}", Plain.paint(&non_copy));
    println!("{}", Plain.paint(&copy));
    // non_copy is still usable here...
    // copy is still usable here...

    println!("{}", Plain.paint(non_copy));
    println!("{}", Plain.paint(copy));
    // non_copy was moved into paint, so it not usable anymore...
    // copy is still usable here...
}

Structs

Painted
Saves a style and a reference to something that will be printed in that style. That something of type T needs to implement either std::fmt::Debug or std::fmt::Display

Style
Saves all properties of a style. Implements ToStyle, so you can call style modifiers on it.

Enums

Attr
Lists possible attributes. It implements ToStyle so it's possible to call ToStyle's methods directly on a Attr variant like:

Color
Lists all possible Colors. It implements ToStyle so it's possible to call ToStyle's methods direclty on a Color variant like:

Traits

ToStyle
Everything that can be seen as part of a style. This is the core of this crate. All functions ("style modifier") consume self and return a modified version of the style.