Struct term_painter::Painted

pub struct Painted<T> {
    // some fields omitted
}

Saves a style and a reference to something that will be printed in that style. That something of type T needs to implement either std::fmt::Debug or std::fmt::Display