Trait term_painter::ToStyle

pub trait ToStyle: Clone {
    fn to_style(self) -> Style;

    fn fg(self, c: Color) -> Style { ... }
    fn bg(self, c: Color) -> Style { ... }
    fn bold(self) -> Style { ... }
    fn dim(self) -> Style { ... }
    fn underline(self) -> Style { ... }
    fn blink(self) -> Style { ... }
    fn reverse(self) -> Style { ... }
    fn secure(self) -> Style { ... }
    fn paint<T>(&self, obj: T) -> Painted<T> { ... }
}

Everything that can be seen as part of a style. This is the core of this crate. All functions ("style modifier") consume self and return a modified version of the style.

Required Methods

fn to_style(self) -> Style

Provided Methods

fn fg(self, c: Color) -> Style

Sets the foreground (text) color.

fn bg(self, c: Color) -> Style

Sets the background color.

fn bold(self) -> Style

Makes the text bold.

fn dim(self) -> Style

Dim mode.

fn underline(self) -> Style

Underlines the text.

fn blink(self) -> Style

Underlines the text.

fn reverse(self) -> Style

Underlines the text.

fn secure(self) -> Style

Secure mode.

fn paint<T>(&self, obj: T) -> Painted<T>

Wraps the style specified in self and something of arbitrary type into a Painted. When Painted is printed it will print the arbitrary something with the given style. T needs to implement std::fmt::Display or std::fmt::Debug.