Trait term_painter::ToStyle

pub trait ToStyle: Sized {
    pub fn to_style(self) -> Style;

    pub fn to_mapped_style<F>(self, func: F) -> Style
    where
        F: FnOnce(&mut Style),
    { ... }

    pub fn fg(self, c: Color) -> Style { ... }

    pub fn bg(self, c: Color) -> Style { ... }

    pub fn bold(self) -> Style { ... }

    pub fn dim(self) -> Style { ... }

    pub fn underline(self) -> Style { ... }

    pub fn not_underline(self) -> Style { ... }

    pub fn blink(self) -> Style { ... }

    pub fn reverse(self) -> Style { ... }

    pub fn secure(self) -> Style { ... }

    pub fn paint<T>(&self, obj: T) -> Painted<T>
    where
        Self: Clone,
    { ... }

    pub fn with<F, R>(&self, f: F) -> R
    where
        F: FnOnce() -> R,
        Self: Clone,
    { ... }
}

Everything that can be seen as part of a style. This is the core of this crate. All functions ("style modifier") consume self and return a modified version of the style.

Required methods

pub fn to_style(self) -> Style

Loading content...
Provided methods

pub fn to_mapped_style<F>(self, func: F) -> Style where
F: FnOnce(&mut Style),

Convenience method for modifying the style before it's returned.

pub fn fg(self, c: Color) -> Style

Sets the foreground (text) color.

pub fn bg(self, c: Color) -> Style

Sets the background color.

pub fn bold(self) -> Style

Makes the text bold.

pub fn dim(self) -> Style

Dim mode.

pub fn underline(self) -> Style

Underlines the text.

pub fn not_underline(self) -> Style

Removes underline-attribute.

pub fn blink(self) -> Style

Underlines the text.

pub fn reverse(self) -> Style

Underlines the text.

pub fn secure(self) -> Style

Secure mode.

pub fn paint<T>(&self, obj: T) -> Painted<T> where
Self: Clone,

Wraps the style specified in self and something of arbitrary type into a Painted. When Painted is printed it will print the arbitrary something with the given style.

pub fn with<F, R>(&self, f: F) -> R where
F: FnOnce() -> R,
Self: Clone,

Executes the given function, applying the style information before calling it and resetting after it finished.

Loading content...